backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1

faster-whisper
groq==0.4.2
requests==2.31.0

pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

spacy==3.7.4
numpy==1.26.4

backboard-sdk
presidio-analyzer
presidio-anonymizer
